Added the side menu panel
What to work on: hide the = button when the side menu is toggled. But fix the formatting.

Work on trying to hide certain timetable blocks when the user clicks on a timetable block in the side menu. It should also show that it has been hidden too on the side menu.