Massive Componentization
Placed all the section colors and timetable parsing from the Timetables controller to the Generator service

Componentized the TimetableViewer. Also made adjustments to the UI of the preferences/restrictions panel.

What to work on:
Make the restrictions/preferences options work in the backend
Allow users to add additional courses
Componentize the homepage
Document the latest changes