Fixed the homepage where it is aligned with the center of the browser. Also made the font of the course search bar bigger
Issue: The course webpage title is not in the middle. Also, the course search bar was too small to be seen in the mobile version of the webpage.

Fix: Made the title of the homepage in the middle as well as increasing the font size of the course search bar.