Refactored the code so that the algorithms and data structures all implement a common interface
Issue #1: The SeasonalTimetable is too coupled with the RedBlackTree class.

Fix #1: Make an interface called IOrderedCollection which the RedBlackTree implements from. Also change the implementation of the SeasonalTimetable where its operations and calculations are based on a generic IOrderedCollection interface.

Issue #2: The root node does not have the "Black" property

Fix #2: The black property is set for each creation of a new RedBlackTree.